+++ src/reducers/containers.js
@@ -1,0 +1,53 @@
+import * as types from '../types';
+import { getRootContainer, updateRootContainer } from '../actions';
+
+function updatedRootContainerState(state = {}, action) {
+  let updatedRootContainer;
+
+  switch (action.type) {
+    case types.REQUEST_ROOT_CONTAINER:
+      return Object.assign({}, state, {
+        isRequesting: true
+      });
+    case types.RECEIVE_ROOT_CONTAINER:
+      // This needs to be removed and a new Action added
+      updatedRootContainer = getRootContainer(action.rootContainer);
+      return Object.assign({}, state, {
+        isRequesting: false,
+        rootContainer: updatedRootContainer,
+        lastUpdated: action.receivedAt
+      });
+    case types.REQUEST_CONTAINER:
+      return Object.assign({}, state, {
+        isRequesting: true
+      });
+    case types.RECEIVE_CONTAINER:
+      // This needs to be removed and a new Action added
+      updatedRootContainer = updateRootContainer(action.container);
+      return Object.assign({}, state, {
+        isRequesting: false,
+        rootContainer: updatedRootContainer,
+        lastUpdated: action.receivedAt
+      });
+    default:
+      return state;
+  }
+}
+
+export function rootContainer(currentState = {}, action) {
+  const initialState = {
+    rootContainer: {},
+    isRequesting: false
+  }
+
+  const state = Object.assign({}, initialState, currentState);
+  const updated = updatedRootContainerState(state.providers, action);
+
+  switch (action.type) {
+    case types.REQUEST_ROOT_CONTAINER:
+    case types.RECEIVE_ROOT_CONTAINER:
+      return Object.assign({}, state, updated);
+    default:
+      return state;
+  }
+}
